refactors database view
